requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
python-dotenv==1.0.1
PyPDF2==3.0.1
python-docx==1.1.2
pinecone-client==5.0.1
requests==2.32.3
